+++ src/test/converter2/issues/gh3012.ts
@@ -1,0 +1,12 @@
+/**
+ * @remarks DictRemarks
+ */
+export const DictionarySchema = {};
+
+/**
+ * {@inheritDoc DictionarySchema}
+ *
+ * @remarks
+ * Alias of {@link DictionarySchema}
+ */
+export const NullProtoObjectSchema = DictionarySchema;